# src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hackaton Prod API",
    "description": "Backend API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/v1/events": {
      "get": {
        "tags": ["events"],
        "summary": "List all events",
        "operationId": "listEvents",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["events"],
        "summary": "Create event",
        "operationId": "createEvent",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              },
              "example": {
                "title": "Tech Conference 2026",
                "date": "2026-06-15T10:00:00.000Z",
                "description": "Annual networking event for developers",
                "tags": ["frontend", "backend", "AI"],
                "isPrivate": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/events/slug/{slug}": {
      "get": {
        "tags": ["events"],
        "summary": "Get event by join link code",
        "operationId": "getEventBySlug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Event link code (format: abc-defg-hij)",
            "schema": {
              "type": "string",
              "pattern": "^[a-z]{3}-[a-z]{4}-[a-z]{3}$",
              "example": "abc-defg-hij"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestSlug"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
    "/v1/events/slug/{slug}/register": {
      "post": {
        "tags": ["events"],
        "summary": "Register for event by join link code",
        "operationId": "registerForEventBySlug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z]{3}-[a-z]{4}-[a-z]{3}$",
              "example": "abc-defg-hij"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Registered for event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRegistrationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestSlug"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
    "/v1/events/{id}": {
      "get": {
        "tags": ["events"],
        "summary": "Get event by id",
        "operationId": "getEventById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      },
      "put": {
        "tags": ["events"],
        "summary": "Update event",
        "description": "Requires event ownership",
        "operationId": "updateEvent",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenEventOwner"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
    "/v1/events/{id}/register": {
      "post": {
        "tags": ["events"],
        "summary": "Register for event",
        "operationId": "registerForEvent",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Registered for event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRegistrationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
    "/v1/events/{id}/participants": {
      "get": {
        "tags": ["events"],
        "summary": "List event participants",
        "operationId": "listEventParticipants",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "description": "Comma-separated tag names. Participant must have all listed tags.",
            "schema": {
              "type": "string",
              "example": "frontend,backend"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of participants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventParticipantResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenNotRegistered"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
    "/v1/events/{id}/stats": {
      "get": {
        "tags": ["events"],
        "summary": "Get event statistics for organizer",
        "description": "Requires event ownership",
        "operationId": "getEventStats",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenEventOwner"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
    "/v1/events/{id}/image": {
      "post": {
        "tags": ["events"],
        "summary": "Upload event cover image",
        "description": "Requires event ownership",
        "operationId": "uploadEventCoverImage",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "slug": "abc-defg-hij",
                  "joinUrl": "https://meetpoint.app/e/abc-defg-hij",
                  "title": "Tech Conference 2026",
                  "date": "2026-06-15T10:00:00.000Z",
                  "description": "Annual networking event for developers",
                  "imageUrl": "https://res.cloudinary.com/demo/image/upload/prod_images/cover.jpg",
                  "organizer": {
                    "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                    "fullName": "Jane Organizer"
                  },
                  "tags": [
                    {
                      "id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
                      "name": "frontend"
                    }
                  ],
                  "createdAt": "2026-05-15T12:00:00.000Z",
                  "updatedAt": "2026-05-16T08:00:00.000Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestImage"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenEventOwner"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundEvent"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "400 — validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError400Validation" }
          }
        }
      },
      "BadRequestImage": {
        "description": "400 — invalid or missing image file",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError400Image" },
            "examples": {
              "missingFile": {
                "summary": "File not sent",
                "value": {
                  "statusCode": 400,
                  "message": "File is required",
                  "error": "Bad Request"
                }
              },
              "invalidType": {
                "summary": "Not an image",
                "value": {
                  "statusCode": 400,
                  "message": "Only image files are allowed",
                  "error": "Bad Request"
                }
              }
            }
          }
        }
      },
      "BadRequestSlug": {
        "description": "400 — invalid event link code",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError400Slug" }
          }
        }
      },
...
      "Unauthorized": {
        "description": "401 — missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError401" }
          }
        }
      },
...
      "ForbiddenEventOwner": {
        "description": "403 — not the event owner",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError403" }
          }
        }
      },
      "ForbiddenNotRegistered": {
        "description": "403 — user is not registered for this event",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError403NotRegistered" }
          }
        }
      },
...
      "NotFoundEvent": {
        "description": "404 — event not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/HttpError404Event" }
          }
        }
      },
...
    },
    "schemas": {
...
      "Tag": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "frontend"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["statusCode", "message"],
        "properties": {
          "statusCode": { "type": "integer" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "error": { "type": "string" }
        }
      },
      "HttpError400Validation": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 400,
          "message": ["title must be longer than or equal to 1 characters"],
          "error": "Bad Request"
        }
      },
      "HttpError400Image": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 400,
          "message": "Only image files are allowed",
          "error": "Bad Request"
        }
      },
      "HttpError400Slug": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 400,
          "message": "Invalid event link code",
          "error": "Bad Request"
        }
      },
...
      "HttpError401": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 401,
          "message": "Unauthorized",
          "error": "Unauthorized"
        }
      },
...
      "HttpError403": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 403,
          "message": "Forbidden",
          "error": "Forbidden"
        }
      },
      "HttpError403NotRegistered": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 403,
          "message": "User is not registered for this event",
          "error": "Forbidden"
        }
      },
...
      "HttpError404Event": {
        "allOf": [{ "$ref": "#/components/schemas/ErrorResponse" }],
        "example": {
          "statusCode": 404,
          "message": "Event not found",
          "error": "Not Found"
        }
      },
...
      "EventParticipantResponse": {
        "type": "object",
        "required": ["userId", "fullName", "tags", "registeredAt"],
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "fullName": { "type": "string", "example": "Alice Smith" },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" }
          },
          "registeredAt": { "type": "string", "format": "date-time" }
        }
      },
      "EventStatsResponse": {
        "type": "object",
        "required": [
          "eventId",
          "participantsRegistered",
          "matchRequestsSent",
          "matchRequestsAccepted",
          "acquaintancesMade"
        ],
        "properties": {
          "eventId": { "type": "string", "format": "uuid" },
          "participantsRegistered": { "type": "integer", "example": 42 },
          "matchRequestsSent": { "type": "integer", "example": 18 },
          "matchRequestsAccepted": { "type": "integer", "example": 7 },
          "acquaintancesMade": { "type": "integer", "example": 7 }
        }
      },
...
      "ImageUploadRequest": {
        "type": "object",
        "required": ["file"],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "example": {
          "file": "photo.jpg"
        }
      },
...
      "EventTag": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "frontend"
          }
        }
      },
      "EventOrganizer": {
        "type": "object",
        "required": ["id", "fullName"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Organizer"
          }
        }
      },
      "EventResponse": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "joinUrl",
          "title",
          "date",
          "description",
          "organizer",
          "tags",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "abc-defg-hij"
          },
          "joinUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://meetpoint.app/e/abc-defg-hij"
          },
          "title": {
            "type": "string",
            "example": "Tech Conference 2026"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-15T10:00:00.000Z"
          },
          "description": {
            "type": "string",
            "example": "Annual networking event for developers"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://res.cloudinary.com/demo/image/upload/prod_images/cover.jpg"
          },
          "organizer": {
            "$ref": "#/components/schemas/EventOrganizer"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventTag"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateEventRequest": {
        "type": "object",
        "required": ["title", "date", "description", "tags"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Tech Conference 2026"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-15T10:00:00.000Z"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "Annual networking event for developers"
          },
          "tags": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "minLength": 1
            },
            "example": ["frontend", "backend", "AI"]
          },
          "isPrivate": {
            "type": "boolean",
            "default": false,
            "example": false,
            "description": "If true, the event is hidden from public listings"
          }
        }
      },
      "UpdateEventRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Tech Conference 2026"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-15T10:00:00.000Z"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "example": "Annual networking event for developers"
          },
          "tags": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "minLength": 1
            },
            "example": ["frontend", "backend"]
          }
        }
      },
      "EventRegistrationResponse": {
        "type": "object",
        "required": ["id", "eventId", "userId", "registeredAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}